+++ packages/nhsuk-frontend/src/nhsuk/components/radios/radios.jsdom.test.mjs
@@ -92,7 +92,7 @@ describe('Radios', () => {
       </form>
     `
 
-    $root = document.querySelector('.nhsuk-radios')
+    $root = document.querySelector(`[data-module="${Radios.moduleName}"]`)
 
     $conditionals = [...$root.querySelectorAll('.nhsuk-radios__conditional')]
 
@@ -131,7 +131,7 @@ describe('Radios', () => {
       $conditionals[0].remove()
 
       expect(() => initRadios()).toThrow(
-        `Radios: Conditional reveal (\`id="${$conditionals[0].id}"\`) not found`
+        `${Radios.moduleName}: Conditional reveal (\`id="${$conditionals[0].id}"\`) not found`
       )
     })
 
@@ -141,7 +141,7 @@ describe('Radios', () => {
       }
 
       expect(() => initRadios()).toThrow(
-        'Radios: Form inputs (`<input type="radio">`) not found'
+        `${Radios.moduleName}: Form inputs (\`<input type="radio">\`) not found`
       )
     })
 
@@ -179,15 +179,15 @@ describe('Radios', () => {
 
     it('should throw with missing $root element', () => {
       expect(() => new Radios()).toThrow(
-        'Radios: Root element (`$root`) not found'
+        `${Radios.moduleName}: Root element (\`$root\`) not found`
       )
     })
 
     it('should throw with wrong $root element type', () => {
-      $root = document.createElementNS('http://www.w3.org/2000/svg', 'svg')
+      const $svg = document.createElementNS('http://www.w3.org/2000/svg', 'svg')
 
-      expect(() => new Radios($root)).toThrow(
-        'Radios: Root element (`$root`) is not of type HTMLElement'
+      expect(() => new Radios($svg)).toThrow(
+        `${Radios.moduleName}: Root element (\`$root\`) is not of type HTMLElement`
       )
     })
   })